+++ app/Http/Controllers/RoleController.php
@@ -6,6 +6,7 @@ use App\Models\Role;
 use App\Http\Requests\StoreRoleRequest;
 use App\Http\Requests\UpdateRoleRequest;
 use App\Models\Permission;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -23,6 +24,10 @@ class RoleController extends Controller
      */
     public function create()
     {
+        if (!Gate::allows('permission', 'create_role')) {
+            return "You do have the permission to access";
+        }
+
         $permissions = Permission::all();
         return view('roles.create', compact('permissions'));
     }
@@ -32,6 +37,10 @@ class RoleController extends Controller
      */
     public function store(StoreRoleRequest $request)
     {
+        if (!Gate::allows('permission', 'create_role')) {
+            return "You do have the permission to access";
+        }
+
         $role = new Role();
         $role->name = $request->name;
         $role->save();
@@ -54,6 +63,10 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        if (!Gate::allows('permission', 'update_role')) {
+            return "You do have the permission to access";
+        }
+
         $permissions = Permission::all();
         $old_permissions = $role->permissions->pluck('id')->toArray();
         return view('roles.edit', compact('role', 'permissions', 'old_permissions'));
@@ -64,6 +77,10 @@ class RoleController extends Controller
      */
     public function update(UpdateRoleRequest $request, Role $role)
     {
+        if (!Gate::allows('permission', 'update_role')) {
+            return "You do have the permission to access";
+        }
+
         $role->name = $request->name;
         $role->update();
         $role->permissions()->sync($request->permission_ids);
@@ -76,6 +93,10 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        if (!Gate::allows('permission', 'delete_role')) {
+            return "You do have the permission to access";
+        }
+
         $role->delete();
         $role->permissions()->detach();
